[sc-58] store local data store local diary entries with either local food id or food id from the database, save locally the food if logging from nutritionix to create later along with the diary entry, decrease request timeout to 2 seconds, add synchronously to diary if added locally, save locally on connection error

+++ lib/models/collection/add_diary_entry_request.dart
@@ -9,7 +9,7 @@ class AddDiaryEntryRequest {
   final String userId;
   final String unitId;
   final String foodId;
-  final int servingQuantity;
+  final double servingQuantity;
   final Meal meal;
 
   const AddDiaryEntryRequest({
